refactor: inline card form container into page
