+++ src/pages/card-form/ui/CardFormPage.tsx
@@ -1,7 +1,29 @@
 import type * as React from "react";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 
-import { CardFormContainer } from "./CardFormContainer";
+import { AppLayout } from "@/widgets/app-layout";
+import { RouteNotFound } from "@/widgets/route-not-found";
+
+import { useCardFormState } from "../model/useCardFormState";
+import { CardEditor } from "./CardEditor";
+
+const CardFormContent: React.FC<{ cardId: string }> = ({ cardId }) => {
+  const navigate = useNavigate();
+  const goBack = () => void navigate(-1);
+  const editor = useCardFormState({ cardId, onCancel: goBack, onSaved: goBack });
+
+  if (editor == null) {
+    return (
+      <RouteNotFound title="Card not found" description="The requested card is unavailable or has been removed." />
+    );
+  }
+
+  return (
+    <AppLayout showHeader>
+      <CardEditor form={editor.form} saveError={editor.saveError} />
+    </AppLayout>
+  );
+};
 
 export const CardFormPage: React.FC = () => {
   const params = useParams();
@@ -9,5 +31,5 @@ export const CardFormPage: React.FC = () => {
   if (cardId == null) throw new Error("invalid card id");
 
   // Form state belongs to one route Card and must reset when the id changes.
-  return <CardFormContainer key={cardId} cardId={cardId} />;
+  return <CardFormContent key={cardId} cardId={cardId} />;
 };
